Integrate duq and duq_sep_bitops
