+++ quantization/learn_both_quantizer_bitwidth/functions/duq.py
@@ -7,6 +7,7 @@ import torch
 import torch.nn as nn 
 import torch.nn.functional as F
 from torch.nn.parameter import Parameter
+from torch import Tensor
 import numpy as np 
 from collections import OrderedDict
 
@@ -25,49 +26,43 @@ class Q_ReLU(nn.Module):
     def __init__(self, act_func=True, inplace=False):
         super(Q_ReLU, self).__init__()
         self.n_lvs = 0
+        self.bits = Parameter(Tensor([32]), requires_grad=False)
         self.act_func = act_func
         self.inplace = inplace
-        self.a = Parameter(torch.Tensor(1))
-        self.c = Parameter(torch.Tensor(1))
-        self.theta_x = Parameter(torch.Tensor(0))
+        self.a = Parameter(Tensor(1))
+        self.c = Parameter(Tensor(1))
+        self.theta = Parameter(Tensor(0))
 
-    def initialize(self, n_lvs, offset, diff, ):
-        self.n_lvs = torch.Tensor(n_lvs) if len(n_lvs)>1 \
-                     else n_lvs[0]
-        self.theta_x = Parameter(torch.ones(len(n_lvs))/len(n_lvs))
+    def initialize(self, bits, offset, diff):
+        self.bits = Parameter(Tensor(bits), requires_grad=False)
+        self.n_lvs = 2 ** self.bits
+
+        self.theta = Parameter(torch.ones(len(self.n_lvs))/len(self.n_lvs))
         self.a.data.fill_(np.log(np.exp(offset + diff)-1))
         self.c.data.fill_(np.log(np.exp(offset + diff)-1))
     
-    def forward(self, x):           
+    def forward(self, x):
         if self.act_func:
             x = F.relu(x, self.inplace)
         
-        if isinstance(self.n_lvs, int) and self.n_lvs == 0:
+        if torch.numel(self.bits)==1 and self.bits == 32:
             return x
         else:
             a = F.softplus(self.a)
             c = F.softplus(self.c)
             x = F.hardtanh(x / a, 0, 1)
             
-            if not isinstance(self.n_lvs, torch.Tensor):
-                x = RoundQuant.apply(x, self.n_lvs) * c
+            if len(self.n_lvs) == 1:
+                x = RoundQuant.apply(x, self.n_lvs[0]) * c
                 return x
             else:
                 # 1) for loop
-                softmask = F.gumbel_softmax(self.theta_x, tau=1, hard=False)
+                softmask = F.gumbel_softmax(self.theta, tau=1, hard=False)
                 softmask = softmask
                 x_bar = torch.zeros_like(x)
-
                 for i, n_lv in enumerate(self.n_lvs):
                     x_bar += RoundQuant.apply(x, n_lv) * c * softmask[i]
-                
-                '''
-                # 2) batched -> out of memory
-                x = x.repeat(self.n_lvs.shape[0], 1, 1, 1, 1)
-                x = RoundQuant.apply(x, self.n_lvs) * c
-                x = softmask * x
-                x = x.sum(dim=0)'''
-                
+                    #x_bar = torch.add(x_bar, RoundQuant.apply(x, n_lv) * c * softmask[i])
                 return x_bar
 
         
@@ -75,50 +70,54 @@ class Q_ReLU6(Q_ReLU):
     def __init__(self, act_func=True, inplace=False):
         super(Q_ReLU6, self).__init__(act_func, inplace)
 
-    def initialize(self, n_lvs, offset, diff):
-        self.n_lvs = torch.Tensor(n_lvs) if len(n_lvs)>1 \
-                     else n_lvs[0]
-        self.theta_x = Parameter(torch.ones(len(n_lvs))/len(n_lvs))
+    def initialize(self, bits, offset, diff):
+        self.bits = Parameter(Tensor(bits), requires_grad=False)
+        self.n_lvs = 2 ** self.bits
+
+        self.theta = Parameter(torch.ones(len(self.n_lvs))/len(self.n_lvs))
         if offset + diff > 6:
             self.a.data.fill_(np.log(np.exp(6)-1))
             self.c.data.fill_(np.log(np.exp(6)-1))
         else:
             self.a.data.fill_(np.log(np.exp(offset + diff)-1))
             self.c.data.fill_(np.log(np.exp(offset + diff)-1))
-
+        
 
 class Q_Sym(nn.Module):
     def __init__(self):
         super(Q_Sym, self).__init__()
         self.n_lvs = 0
-        self.a = Parameter(torch.Tensor(1))
-        self.c = Parameter(torch.Tensor(1))
-        self.theta_x = Parameter(torch.Tensor(0))
+        self.bits = Parameter(Tensor([32]), requires_grad=False)
+        self.a = Parameter(Tensor(1))
+        self.c = Parameter(Tensor(1))
+        self.theta = Parameter(Tensor(0))
 
-    def initialize(self, n_lvs, offset, diff):
-        self.n_lvs = torch.Tensor(n_lvs) if len(n_lvs)>1 \
-                     else n_lvs[0]
+    def initialize(self, bits, offset, diff):
+        self.bits = Parameter(Tensor(bits), requires_grad=False)
+        self.n_lvs = 2 ** self.bits
+
+        self.theta = Parameter(torch.ones(len(self.n_lvs))/len(self.n_lvs))
         self.a.data.fill_(np.log(np.exp(offset + diff)-1))
         self.c.data.fill_(np.log(np.exp(offset + diff)-1))
-        self.theta_x = Parameter(torch.ones(len(n_lvs))/len(n_lvs))
 
     def forward(self, x):
-        if isinstance(self.n_lvs, int) and self.n_lvs == 0:
+        if torch.numel(self.bits)==1 and self.bits == 32:
             return x
         else:
             a = F.softplus(self.a)
             c = F.softplus(self.c)
             x = F.hardtanh(x / a, -1, 1)
 
-            if not isinstance(self.n_lvs, torch.Tensor):
-                x = RoundQuant.apply(x, self.n_lvs // 2) * c
+            if len(self.n_lvs) == 1:
+                x = RoundQuant.apply(x, self.n_lvs[0] // 2) * c
                 return x
             else:
-                softmask = F.gumbel_softmax(self.theta_x, tau=1, hard=False)
+                softmask = F.gumbel_softmax(self.theta, tau=1, hard=False)
                 softmask = softmask
                 x_bar = torch.zeros_like(x)
                 for i, n_lv in enumerate(self.n_lvs):
                     x_bar += RoundQuant.apply(x, n_lv) * c * softmask[i]
+                    #x_bar = torch.add(x_bar, RoundQuant.apply(x, n_lv) * c * softmask[i])
                 return x_bar
 
 
@@ -126,10 +125,11 @@ class Q_HSwish(nn.Module):
     def __init__(self, act_func=True):
         super(Q_HSwish, self).__init__()
         self.n_lvs = 0
+        self.bits = Parameter(Tensor([32]), requires_grad=False)
         self.act_func = act_func
-        self.a = Parameter(torch.Tensor(1))
+        self.a = Parameter(Tensor(1))
         self.b = 3/8
-        self.c = Parameter(torch.Tensor(1))
+        self.c = Parameter(Tensor(1))
         self.d = -3/8
 
     def initialize(self, n_lvs, offset, diff):
@@ -141,7 +141,7 @@ class Q_HSwish(nn.Module):
         if self.act_func:
             x = x * (F.hardtanh(x + 3, 0, 6) / 6)
 
-        if self.n_lvs == 0:
+        if torch.numel(self.bits)==1 and self.bits == 32:
             return x
         else:
             a = F.softplus(self.a)
@@ -157,49 +157,41 @@ class Q_Conv2d(nn.Conv2d):
     def __init__(self, *args, **kargs):
         super(Q_Conv2d, self).__init__(*args, **kargs)
         self.n_lvs = 0
-        self.a = Parameter(torch.Tensor(1))
-        self.c = Parameter(torch.Tensor(1))
+        self.bits = Parameter(Tensor([32]), requires_grad=False)
+        self.a = Parameter(Tensor(1))
+        self.c = Parameter(Tensor(1))
         self.weight_old = None
-        self.theta_w = Parameter(torch.Tensor(0))
+        self.theta = Parameter(Tensor(0))
 
-    def initialize(self, n_lvs):
-        self.n_lvs = torch.Tensor(n_lvs).view(-1,1,1,1,1) if len(n_lvs)>1 \
-                     else n_lvs[0]
-        self.theta_w = Parameter(torch.ones(len(n_lvs))/len(n_lvs))
+    def initialize(self, bits):
+        self.bits = Parameter(Tensor(bits), requires_grad=False)
+        self.n_lvs = 2 ** self.bits
+        
+        self.theta = Parameter(torch.ones(len(self.n_lvs))/len(self.n_lvs))
         max_val = self.weight.data.abs().max().item()
         self.a.data.fill_(np.log(np.exp(max_val * 0.9)-1))
         self.c.data.fill_(np.log(np.exp(max_val * 0.9)-1))
-        
 
     def _weight_quant(self):
         a = F.softplus(self.a)
         c = F.softplus(self.c)
         weight = F.hardtanh(self.weight / a, -1, 1)
 
-        if not isinstance(self.n_lvs, torch.Tensor):
-            weight = RoundQuant.apply(weight, self.n_lvs // 2) * c
-            bitwidth = self.n_lvs
+        if len(self.n_lvs) == 1:
+            weight = RoundQuant.apply(weight, self.n_lvs[0] // 2) * c
+            return weight
         else:
-            softmask = F.gumbel_softmax(self.theta_w, tau=1, hard=False)
-            softmask = softmask.view(-1,1,1,1,1)
-
-            # 1) batched -> out of memory
-            weight = weight.repeat(self.n_lvs.shape[0], 1, 1, 1, 1)
-            weight = RoundQuant.apply(weight, self.n_lvs) * c
-            weight = softmask * weight
-            weight = weight.sum(dim=0)
-
-            '''# 2) for loop
+            softmask = F.gumbel_softmax(self.theta, tau=1, hard=False)
+            softmask = softmask.view(-1,1,1,1,1)       
             w_bar = torch.zeros_like(weight)
             for i, n_lv in enumerate(self.n_lvs):
-                w_bar += RoundQuant.apply(weight, n_lv) * c * softmask[i,0,0,0,0]'''
+                w_bar += RoundQuant.apply(weight, n_lv) * c * softmask[i,0,0,0,0]
+                #w_bar = torch.add(w_bar, RoundQuant.apply(weight, n_lv) * c * softmask[i,0,0,0,0])
 
-
-        return weight
+            return w_bar
 
     def forward(self, x):
-        O, I, K1, K2 = self.weight.shape
-        if isinstance(self.n_lvs, int) and self.n_lvs == 0:
+        if torch.numel(self.bits)==1 and self.bits == 32:
             return F.conv2d(x, self.weight, self.bias,
                 self.stride, self.padding, self.dilation, self.groups)
         else:
@@ -212,14 +204,16 @@ class Q_Linear(nn.Linear):
     def __init__(self, *args, **kargs):
         super(Q_Linear, self).__init__(*args, **kargs)
         self.n_lvs = 0
-        self.a = Parameter(torch.Tensor(1))
-        self.c = Parameter(torch.Tensor(1))
+        self.bits = Parameter(Tensor([32]), requires_grad=False)
+        self.a = Parameter(Tensor(1))
+        self.c = Parameter(Tensor(1))
         self.weight_old = None
 
-    def initialize(self, n_lvs):
-        self.n_lvs = torch.Tensor(n_lvs).view(-1,1,1) if len(n_lvs)>1 \
-                     else n_lvs[0]
-        self.theta_w = Parameter(torch.ones(len(n_lvs))/len(n_lvs))
+    def initialize(self, bits):
+        self.bits = Parameter(Tensor(bits), requires_grad=False)
+        self.n_lvs = 2 ** self.bits
+
+        self.theta = Parameter(torch.ones(len(self.n_lvs))/len(self.n_lvs))
         max_val = self.weight.data.abs().max().item()
         self.a.data.fill_(np.log(np.exp(max_val * 0.9)-1))
         self.c.data.fill_(np.log(np.exp(max_val * 0.9)-1))
@@ -229,31 +223,25 @@ class Q_Linear(nn.Linear):
         c = F.softplus(self.c)
 
         weight = F.hardtanh(self.weight / a, -1, 1)
-        if not isinstance(self.n_lvs, torch.Tensor):
-            weight = RoundQuant.apply(weight, self.n_lvs // 2) * c
-            bitwidth = self.n_lvs
+        if len(self.n_lvs) == 1:
+            weight = RoundQuant.apply(weight, self.n_lvs[0] // 2) * c
+            return weight
         else:
-            softmask = F.gumbel_softmax(self.theta_w, tau=1, hard=False)
+            softmask = F.gumbel_softmax(self.theta, tau=1, hard=False)
             softmask = softmask.view(-1,1,1)
-            weight = weight.repeat(self.n_lvs.shape[0], 1, 1)
-            weight = RoundQuant.apply(weight, self.n_lvs) * c
-            weight = softmask * weight
-            weight = weight.sum(dim=0)
 
-            '''# 2) for loop
             w_bar = torch.zeros_like(weight)
             for i, n_lv in enumerate(self.n_lvs):
-                w_bar += RoundQuant.apply(weight, n_lv) * c * softmask[i,0,0]'''
-            bitwidth = (softmask * self.n_lvs).sum()
+                w_bar += RoundQuant.apply(weight, n_lv) * c * softmask[i,0,0]
+                #w_bar = torch.add(w_bar, RoundQuant.apply(weight, n_lv) * c * softmask[i,0,0])
 
-        return weight, bitwidth
+            return w_bar
 
     def forward(self, x):
-        O, I = self.weight.shape
-        if isinstance(self.n_lvs, int) and self.n_lvs == 0:
+        if torch.numel(self.bits)==1 and self.bits == 32:
             return F.linear(x, self.weight, self.bias)
         else:
-            weight= self._weight_quant()
+            weight = self._weight_quant()
             return F.linear(x, weight, self.bias)
 
 
@@ -281,10 +269,12 @@ class Q_Conv2dPad(Q_Conv2d):
 
 
 
-def initialize(model, loader, n_lvs, act=False, weight=False, eps=0.05):
+def initialize(model, loader, bits, act=False, weight=False, eps=0.05):
+    if isinstance(bits, int):
+        bits = [bits]
     def initialize_hook(module, input, output):
         if isinstance(module, (Q_ReLU, Q_Sym, Q_HSwish)) and act:
-            if not isinstance(input, torch.Tensor):
+            if not isinstance(input, list):
                 input = input[0]
             input = input.detach().cpu().numpy()
 
@@ -301,12 +291,11 @@ def initialize(model, loader, n_lvs, act=False, weight=False, eps=0.05):
                 small, large = 0, 1e-3
             else:
                 small, large = input[int(len(input) * eps)], input[int(len(input) * (1-eps))]
-
-            module.initialize(n_lvs, small, large - small)
+            module.initialize(bits, small, large - small)
 
         if isinstance(module, (Q_Conv2d, Q_Linear)) and weight:
-            module.initialize(n_lvs)
-
+            module.initialize(bits)
+        
         if isinstance(module, Q_Conv2d):
             O, I, K1, K2 = module.weight.shape
             N, C, H, W = input[0].shape
@@ -324,9 +313,9 @@ def initialize(model, loader, n_lvs, act=False, weight=False, eps=0.05):
         hook = module.register_forward_hook(initialize_hook)
         hooks.append(hook)
 
-
     model.train()
     model.cpu()
+
     for i, (input, target) in enumerate(loader):
         with torch.no_grad():
             if isinstance(model, nn.DataParallel):
